@@ -22,17 +22,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.2",
+    "@angular/animations": "8.2.3",
     "@angular/cdk": "^7.1.1",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
+    "@angular/common": "8.2.3",
+    "@angular/compiler": "8.2.3",
+    "@angular/core": "8.2.3",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "~7.1.0",
+    "@angular/forms": "8.2.3",
     "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/platform-browser": "8.2.3",
+    "@angular/platform-browser-dynamic": "8.2.3",
+    "@angular/router": "8.2.3",
     "@ngrx/effects": "^6.1.2",
     "@ngrx/entity": "^6.1.2",
     "@ngrx/router-store": "^6.1.2",
@@ -42,15 +42,15 @@
     "hammerjs": "^2.0.8",
     "ngrx-store-freeze": "^0.2.4",
     "ngx-toastr": "^9.1.1",
-    "rxjs": "~6.3.3",
+    "rxjs": "6.5.2",
     "tslib": "^1.9.0",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.802.2",
-    "@angular/cli": "~8.2.2",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular/cli": "8.2.2",
+    "@angular/compiler-cli": "8.2.3",
+    "@angular/language-service": "8.2.3",
     "@ngrx/schematics": "^6.1.2",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
@@ -58,18 +58,11 @@
     "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-marbles": "^0.4.1",
-    "jasmine-spec-reporter": "~4.2.1",
     "jest": "^23.6.0",
     "jest-preset-angular": "^6.0.2",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.12.0",
-    "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "3.5.3"
   }
 }
